First ever commit for this project
This is a very early alpha version of piDisplay. This module can display
numbers and some letters on a 7 segment, 2 digit display that is
connected to a Raspberry Pi. It runs via node.js. I don't have time to
go into detail about how all this works at the moment, so you'll have to
look through the code if you're curious about using this module. I will
try to update this repository soon. Thanks for your patience.